#1521 - Created the BirthDateSelector component
